fixing naming issue and size_t

+++ PrefixMatcher.cpp
@@ -10,7 +10,7 @@ RouterNode::~RouterNode() {
 
 void PrefixMatcher::insert(std::string address, int routerNumber) {
   RouterNode* curr = root;
-  for (int i = 0; i < address.length(); i++) {
+  for (size_t i = 0; i < address.length(); i++) {
     if (curr->links[address[i] - '0'] == nullptr) {
       curr->links[address[i] - '0'] = new RouterNode(-1);
     }
@@ -22,7 +22,7 @@ void PrefixMatcher::insert(std::string address, int routerNumber) {
 int PrefixMatcher::selectRouter(std::string networkAddress) {
   RouterNode* curr = root;
   int last_router = -1;
-  int i = 0;
+  size_t i = 0;
   while (curr != nullptr && i < networkAddress.length()) {
     if (curr->routerNumber != -1) last_router = curr->routerNumber;
     curr = curr->links[networkAddress[i] - '0'];
